+++ src/domain/error.rs
@@ -1,0 +1,1 @@
+pub mod kind;
