+++ src/domain/service/job_execution_service.rs
@@ -25,4 +25,3 @@ impl JobExecuteService {
         self.scheduler.add(job).await.unwrap();
     }
 }
-
